This commit includes: -Header guard convention fix. -Added another event type. -Compressed the size of some events slightly. -Added some ToString methods for enums. -Implemented EventHandler functions now automatically update with FEvent constructor modifications. -Added some basic Event debug functionality.

+++ Solution/PhoenixEngine/Source/Platform/Input/InputAction.h
@@ -8,7 +8,7 @@ namespace Phoenix
 {
 	namespace EInputAction
 	{
-		typedef UInt32 Type;
+		typedef UInt8 Type;
 
 		enum Value : Type
 		{
